criado o banco de dados para Estudio. criado o HELPER para Estudio Refatorando código: organizando em pastas.

+++ app/src/main/java/br/com/pos/cacaestudio/adapter/EstudiosAdapter.java
@@ -5,13 +5,12 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import java.util.List;
 
 import br.com.pos.cacaestudio.R;
-import br.com.pos.cacaestudio.modelo.Estudio;
+import br.com.pos.cacaestudio.modelo.entity.Estudio;
 
 /**
  * Created by livia on 17/02/2018.
@@ -61,7 +60,7 @@ public class EstudiosAdapter extends BaseAdapter {
 
         campoNome.setText(estudio.getNome());
         campoEndereco.setText(estudio.getEndereco());
-        campoPreco.setText(estudio.getPreco());
+        campoPreco.setText((int) estudio.getPreco());
 
         return view;
     }
